Remove nonworking interactive.py.
